Add a base class for notifications
